feat(jrs): the index domain of 7.3.18 in the walk of 23.1.3
7.3.18 reads a `length` up to 2^53-1 and every clause of 23.1.3 walks
the indices below it; the walk held its own in `u32`, so a `length` past
2^32-1 was clamped and an index above that was never reached. The state
carries the three as `i64` now, and an index no Elements store could
hold is read as the ordinary name 10.4.2.1 gives it.

Full run at `--fuel 1000000 --all`, no masks: 59,233 passed, 19,655
failed, 24,037 unsupported of 102,925 variants; gained 14, lost 0. The
blockers of the switch fall to 3,512: 106 failures and 3,406 gaps. The
stack path is untouched.

+++ crates/jrs/src/engine/object.rs
@@ -231,12 +231,14 @@ pub enum ObjectKind {
         output: Value,
         /// The element the callback that is in flight was given.
         element: Value,
-        /// Its index, which is the second argument of the callback.
-        element_index: u32,
+        /// Its index, which is the second argument of the callback. 7.3.18
+        /// reads a `length` up to 2^53-1, so an index of the walk is no
+        /// narrower than that.
+        element_index: i64,
         /// The index the next callback is for.
-        index: u32,
+        index: i64,
         /// Where the walk stops, taken once as 23.1.3 takes it.
-        length: u32,
+        length: i64,
         /// Whether an accumulator has been taken yet (23.1.3.24 step 6).
         started: bool,
         /// Whether the call in flight is the getter of an accessor element
